Added infrastructure layer and JsonFile repository

+++ app/domain/entities/Item.php
@@ -31,7 +31,7 @@ class Item
   /**
    * @return int
    */
-  public function getPrice(): int
+  public function getPrice(): float
   {
     return $this->price;
   }
